Updating service to cleanup and freeze
