+++ requirements.txt
@@ -2,3 +2,4 @@ requests>=2.26.0
 beautifulsoup4>=4.10.0
 Flask>=2.0.0
 lxml
+Flask-Caching
